requirements.txt
Pillow
packaging
pyinstaller
flet
requests
